Support @htmlUrl, @description, and @title
NOTE: This modifies how @htmlUrl was previously stored.

These three attributes are supposed to be filled with information
already found in the feed itself. It is possible that they may
be out-of-sync with the feed or erroneously populated, though.
Because they contain duplicate information, they are not treated
as first-class data (additionally, @title would conflict with
feed['title']).