Fix most Jython unit test errors
Jython (currently) raises NullPointerExceptions when getting
XML attributes without a namespace using the attrs[(None, 'attr')]
notation, so I introduced the NONS constant which will be None
on CPython, and '' on Jython.

Jython doesn't catch org.xml.sax.SAXParseExceptions in the
ErrorHandler functions, so it's necessary to import it and
catch the exception outside of the ErrorHandler functions.