Introduce No-Eval and Requests syntax in testfiles
Requests represents the number of HTTP requests that the file will
be involved in. If the filename contains 'http' then Requests
defaults to 1.

No-Eval indicates that no TestCase will be automatically generated
for the file. This is useful for the files that will be the final
destination of an HTTP redirect, or for a file that otherwise
exists solely as an endpoint of some kind.