Add an Event to signal the server thread is ready
Due to a problem with threading, the unit tests would begin running
before the webserver was ready to receive localhost URL requests.
In Python 3 on my machine there would be an error once every 125
runs or so; with this code I'm unable to recreate the problem even
after tens of thousands of runs. I believe the problem is fixed.